# spec/openapi-dashboard.yml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: E2B Dashboard API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminJWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    projectID:
      name: projectID
      in: path
      required: true
      description: Identifier of the project.
      schema:
        type: string
        format: uuid
    userID:
      name: userID
      in: path
      required: true
      description: Identifier of the user.
      schema:
        type: string
        format: uuid
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "409":
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code.
        message:
          type: string
          description: Error message.
...
    ManagementProjectMemberIdentity:
      type: object
      required: [issuer, subject]
      properties:
        issuer:
          type: string
          minLength: 1
          maxLength: 2048
        subject:
          type: string
          minLength: 1
          maxLength: 2048
    ManagementProjectMemberApplyRequest:
      type: object
      required: [revision, present]
      properties:
        revision:
          type: integer
          format: int64
          minimum: 1
        present:
          type: boolean
        identities:
          type: array
          maxItems: 16
          items:
            $ref: "#/components/schemas/ManagementProjectMemberIdentity"
...
paths:
...
  /v1/management/projects/{projectID}/members/{userID}:
    parameters:
      - $ref: "#/components/parameters/projectID"
      - $ref: "#/components/parameters/userID"
    put:
      operationId: managementApplyProjectMember
      summary: Apply one versioned project member projection (v1).
      description: >-
        Applies the newest desired presence for one project member. An older
        or duplicate revision is accepted without changing target state.
      tags: [control-plane-management]
      security:
        - AdminJWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementProjectMemberApplyRequest"
      responses:
        "204":
          description: Membership projection is applied or already superseded.
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
...